+++ packages/terra-time-input/CHANGELOG.md
@@ -6,6 +6,9 @@ Unreleased
 ### Added
 * Added en-AU translations
 
+### Fixed
+* Changed browser.execute function in wdio spec to be a string for IE and FF driver compatibility
+
 3.1.0 - (February 5, 2019)
 ------------------
 ### Changed